requirements.txt
# Core dependencies
httpx==0.27.0
python-dotenv==1.0.1
tqdm==4.66.4

# Async utilities
# asyncio is part of the Python standard library; do not install separately

# Development dependencies (optional)
# pytest==8.0.0
# pytest-asyncio==0.23.5
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0
